Add files via upload
Note that RNASeqProcessing_Github also requires the raw Kallisto outputs as input files. For the purposes of the Figure in this paper and space limitations on Github, the raw sequencing data is not uploaded; however, the processed files are available for inputs to the R code that performs differential gene expression analysis.